refactor(web): move components to their own dir
