+++ packages/web/src/lib/components/AutomatticLogo.svelte
