+++ packages/web/src/lib/components/ChromeLogo.svelte
