+++ packages/web/src/lib/components/CodeLogo.svelte
