+++ packages/web/src/lib/components/DefaultNeovimConfig.svelte
