+++ packages/web/src/lib/components/EdgeLogo.svelte
