+++ packages/web/src/lib/components/Editor.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import LintCard from '$lib/LintCard.svelte';
+import LintCard from '$lib/components/LintCard.svelte';
 import { Card } from 'flowbite-svelte';
 import { type WorkerLinter } from 'harper.js';
 import {
