+++ packages/web/src/lib/components/EmacsLogo.svelte
