+++ packages/web/src/lib/components/FirefoxLogo.svelte
