+++ packages/web/src/lib/components/HelixLogo.svelte
