+++ packages/web/src/lib/components/LintKindChart.svelte
