+++ packages/web/src/lib/components/Logo.svelte
