+++ packages/web/src/lib/components/NeovimLogo.svelte
