+++ packages/web/src/lib/components/ObsidianLogo.svelte
