+++ packages/web/src/lib/components/SublimeLogo.svelte
