+++ packages/web/src/lib/components/WordPressLogo.svelte
