+++ packages/web/src/lib/components/ZedLogo.svelte
