+++ packages/web/src/routes/+layout.svelte
@@ -2,8 +2,8 @@
 import '../app.css';
 
 import { browser } from '$app/environment';
-import AutomatticLogo from '$lib/AutomatticLogo.svelte';
-import GutterCenter from '$lib/GutterCenter.svelte';
+import AutomatticLogo from '$lib/components/AutomatticLogo.svelte';
+import GutterCenter from '$lib/components/GutterCenter.svelte';
 import posthog from 'posthog-js';
 import { onMount } from 'svelte';
